Move CreateHTTPClients to networking.go
